python/pyproject.toml
[project]
name = "openclaw"
version = "0.1.0"
description = "LLM-driven options trading orchestrator — research, thesis, execution"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "pydantic>=2.0",
    "anthropic>=0.40.0",
    "instructor>=1.7.0",
    "yfinance>=0.2.40",
    "apscheduler>=4.0.0a5",
    "httpx>=0.27.0",
    "python-telegram-bot>=21.0",
    "beautifulsoup4>=4.12",
    "feedparser>=6.0",
    "ta>=0.11.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[project.scripts]
openclaw = "openclaw.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
